+++ DataDefinitions/StationModel.cs
@@ -11,25 +11,23 @@ namespace EddiDataDefinitions
             resourceManager = Properties.StationModels.ResourceManager;
             resourceManager.IgnoreCase = true;
             missingEDNameHandler = (edname) => new StationModel(edname);
-
-            None = new StationModel("None");
-            var AsteroidBase = new StationModel("AsteroidBase");
-            var Bernal = new StationModel("Bernal");
-            var Coriolis = new StationModel("Coriolis");
-            var Megaship = new StationModel("Megaship");
-            var MegaShipCivilian = new StationModel("MegaShipCivilian");
-            var Ocellus = new StationModel("Ocellus");
-            var Orbis = new StationModel("Orbis");
-            var Outpost = new StationModel("Outpost");
-            var SurfaceStation = new StationModel("SurfaceStation");
-            var CraterOutpost = new StationModel("CraterOutpost");
-            var CraterPort = new StationModel("CraterPort");
-            var OutpostScientific = new StationModel("OutpostScientific");
-            var FleetCarrier = new StationModel("FleetCarrier");
-            var OnFootSettlement = new StationModel("OnFootSettlement");
         }
 
-        public static readonly StationModel None;
+        public static readonly StationModel None = new StationModel("None");
+        public static readonly StationModel AsteroidBase = new StationModel("AsteroidBase");
+        public static readonly StationModel Bernal = new StationModel("Bernal");
+        public static readonly StationModel Coriolis = new StationModel("Coriolis");
+        public static readonly StationModel Megaship = new StationModel("Megaship");
+        public static readonly StationModel MegaShipCivilian = new StationModel("MegaShipCivilian");
+        public static readonly StationModel Ocellus = new StationModel("Ocellus");
+        public static readonly StationModel Orbis = new StationModel("Orbis");
+        public static readonly StationModel Outpost = new StationModel("Outpost");
+        public static readonly StationModel SurfaceStation = new StationModel("SurfaceStation"); // No longer used in the player journal
+        public static readonly StationModel CraterOutpost = new StationModel("CraterOutpost");
+        public static readonly StationModel CraterPort = new StationModel("CraterPort");
+        public static readonly StationModel OutpostScientific = new StationModel("OutpostScientific");
+        public static readonly StationModel FleetCarrier = new StationModel("FleetCarrier");
+        public static readonly StationModel OnFootSettlement = new StationModel("OnFootSettlement");
 
         // dummy used to ensure that the static constructor has run
         public StationModel() : this("")
@@ -46,39 +44,35 @@ namespace EddiDataDefinitions
             }
 
             // Translate from EDSM / EDDB station model names if these are present
-            Dictionary<string, string> modelTranslations = new Dictionary<string, string>()
+            Dictionary<string, StationModel> modelTranslations = new Dictionary<string, StationModel>()
             {
-                { "Coriolis Starport", "Coriolis" },
-                // Ocellus starports are described by the journal as either "Bernal" or "Ocellus"
-                { "Bernal Starport", "Bernal" },
-                { "Ocellus Starport", "Ocellus" },
-                { "Orbis Starport", "Orbis" },
+                { "Coriolis Starport", Coriolis },
+                { "Bernal Starport", Bernal }, // Ocellus starports are described by the journal as either "Bernal" or "Ocellus"
+                { "Ocellus Starport", Ocellus },
+                { "Orbis Starport", Orbis },
                 // The journal doesn't provide details on the type of outpost (military, civilian, scientific, etc.)
                 // Types are likely derived from the station primary economy, but mixing these into the model name does not add value.
-                { "Outpost", "Outpost" },
-                { "Civilian Outpost", "Outpost" },
-                { "Commercial Outpost", "Outpost" },
-                { "Industrial Outpost", "Outpost" },
-                { "Military Outpost", "Outpost" },
-                { "Mining Outpost", "Outpost" },
-                { "Scientific Outpost", "Outpost" },
-                { "Unknown Outpost", "Outpost" },
-                // Planetary ports are assigned by EDSM via marketID, from a list given to EDSM by FDev long ago.
-                // The journal doesn't distinguish between planetary ports and outposts. 
-                { "Planetary Outpost", "SurfaceStation" },
-                { "Planetary Port", "SurfaceStation" },
-                // Planetary settlements are not dockable and require manual edits on ROSS to add to EDDB
-                { "Planetary Settlement", "SurfaceStation" },
-                { "Planetary Engineer Base", "SurfaceStation" },
-                { "Unknown Planetary", "SurfaceStation" },
-                { "Asteroid base", "AsteroidBase" },
-                { "Mega ship", "Megaship" },
+                { "Civilian Outpost", Outpost },
+                { "Commercial Outpost", Outpost },
+                { "Industrial Outpost", Outpost },
+                { "Military Outpost", Outpost },
+                { "Mining Outpost", Outpost },
+                { "Scientific Outpost", Outpost },
+                { "Unknown Outpost", Outpost },
+                { "Planetary Outpost", CraterOutpost },
+                { "Planetary Port", CraterPort },
+                { "Planetary Settlement", SurfaceStation }, // Planetary settlements are not dockable and require manual edits on ROSS to add to EDDB
+                { "Planetary Engineer Base", CraterOutpost },
+                { "Unknown Planetary", SurfaceStation },
+                { "Asteroid base", AsteroidBase },
+                { "Mega ship", Megaship },
+                { "Odyssey Settlement", OnFootSettlement },
             };
-            foreach (KeyValuePair<string, string> model in modelTranslations)
+            foreach (var model in modelTranslations)
             {
                 if (from == model.Key)
                 {
-                    return FromEDName(model.Value);
+                    return model.Value;
                 }
             }